Updated to read spin-polarised density.
A few other code clean up and changes - the main one in terms of functionality is how we deal with the sum of a complex charge density (previously took the absolute value first and then summed) - now we take the absolute value after the sum (negative charge density from Fourier transform noise with MC). 

Also no longer require inversion symmetry and odd grids as they may not be!